[Feat]-사용자 상태에 따른 페이지 라우팅
