+++ src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import AppRouter from "./routes/Router";
+import AppRouter from "@/router";
 
 function App() {
   return (
